Extract jwt logic to commons
